test: a few template tests for linter service
