+++ src/cli/services/__fixtures__/draft-nested-ref.oas2.json
