+++ src/cli/services/__fixtures__/draft-ref.oas2.json
