+++ src/cli/services/__fixtures__/openapi-3.0-no-contact.yaml
