+++ src/cli/services/__fixtures__/openapi-3.0-valid.yaml
