+++ src/cli/services/__fixtures__/refs/info-ref.json
