+++ src/cli/services/__fixtures__/refs/info.json
