+++ src/cli/services/__fixtures__/ruleset-extends-invalid.yaml
