+++ src/cli/services/__fixtures__/ruleset-extends-valid.yaml
