+++ src/cli/services/__fixtures__/ruleset-invalid.yaml
