+++ src/cli/services/__fixtures__/spectral.yml
